snapshot: add password argument for SSH
To make this tool automatable with no or little system setup, it is
required that it accepts a `password` argument to pass to SSH for the
cases where publickey access is not configured.

This commit adds the `--password` commandline option and uses it along
side `sshpass` in order to facilitate this.