snapshot: fail better on error
In the case where something went wrong with the subcommand, for example
SSH failed, etc., the tool would fail with an unhelpful error because the
success of the command was not checked before the output was passed.

This commit changes this so that the success of the command is checked
for the stats are returned for further processing, printing an error
message and exiting 1 if anything went wrong.

Note that some SSH options are set to prevent clutter in the error message
from SSH.