Fix ssh'ing as non-root user.
You may want to ssh into another machine as non-root. You can't see
all of the info in /proc as non-root usually but can see some.

The issue was that the first tail command can fail if it can't read
certain files, which would prevent the collection of /proc/**/stat
info due to the `&&`. It would also kill progress of the script as
we'd get a non-zero exit from sshpass.

By running the commands one after another the fail exit from the
first tail command is consumed. This is OK as we still get some
info.

Also run both commands under nice, rather than just the first tail
command.