Refactor parsers to have a generic model: stat
This is a next part of a re-factoring attempt to make
parsers generic. It appears that keeping data in a dictionary
will make more sense than having to maintain new types in 'model'.

This no functional change adds a new parser for stat, but still leaves
tail parsing 'as is' (i.e. it's implementation still uses old parser).